refactor(frontend!): Organising files as per their use
