+++ frontend/src/hooks/useProductFilter.js
