+++ frontend/src/utils/truncateText.js
